% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brothers}
\alias{brothers}
\title{Data about brothers' income}
\format{a data frame with 14 rows and 3 variables
\describe{
  \item{family}{identifier for the family}
  \item{older}{income of older brother}
  \item{younger}{income of younger brother}
}}
\source{
\url{https://dahl.byu.edu/223/2018a2/l12/brothers.txt}
}
\usage{
brothers
}
\description{
A dataset containing income details about older and younger brothers.
}
\keyword{datasets}
