% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpa}
\alias{gpa}
\title{GPA data}
\format{a data frame with 1000 rows and 3 variables
\describe{
  \item{sat}{SAT scores}
  \item{hsgpa}{High School GPA}
  \item{cgpa}{College GPA}
}}
\source{
\url{https://dahl.byu.edu/223/2018a2/l12/gpa.txt}
}
\usage{
gpa
}
\description{
A dataset containing SAT score, High School GPA, and college GPA for 1000
students.
}
\keyword{datasets}
